Refactor context providers for lint compliance
