+++ src/hooks/AuthProvider.tsx
@@ -1,0 +1,66 @@
+import { useEffect, useMemo, useState } from 'react';
+import type { ReactNode } from 'react';
+import type { User, Session } from '@supabase/supabase-js';
+import { supabase, clearSupabaseSession } from '@/integrations/supabase';
+import { AuthContext, type AuthContextValue } from './auth-context';
+
+interface AuthProviderProps {
+  children: ReactNode;
+}
+
+/**
+ * Provides Supabase authentication state to descendant components.
+ */
+export function AuthProvider({ children }: AuthProviderProps) {
+  const [user, setUser] = useState<User | null>(null);
+  const [session, setSession] = useState<Session | null>(null);
+  const [isLoading, setIsLoading] = useState(true);
+
+  useEffect(() => {
+    const {
+      data: { subscription },
+    } = supabase.auth.onAuthStateChange((_event, authSession) => {
+      setSession(authSession);
+      setUser(authSession?.user ?? null);
+      setIsLoading(false);
+    });
+
+    supabase.auth.getSession().then(({ data: { session: initialSession } }) => {
+      setSession(initialSession);
+      setUser(initialSession?.user ?? null);
+      setIsLoading(false);
+    });
+
+    return () => subscription.unsubscribe();
+  }, []);
+
+  const signOut = async () => {
+    try {
+      const { error } = await supabase.auth.signOut();
+      if (error) {
+        throw error;
+      }
+    } catch (error) {
+      console.error('Erro ao encerrar sessão', error);
+      throw error;
+    } finally {
+      clearSupabaseSession();
+      setSession(null);
+      setUser(null);
+    }
+  };
+
+  const value = useMemo<AuthContextValue>(
+    () => ({
+      user,
+      session,
+      isLoading,
+      signOut,
+    }),
+    [user, session, isLoading]
+  );
+
+  return <AuthContext.Provider value={value}>{children}</AuthContext.Provider>;
+}
+
+export default AuthProvider;
